src/mmv/mmvskia/requirements.txt
ffmpeg-python==0.2.0
gitpython==3.1.11
glfw==2.0.0
librosa==0.8.0
midi2audio==0.1.1
mido==1.2.9
opencv-python==4.4.0.46
pillow-simd==7.0.0.post3
pip-chill==1.0.0
pyunpack==0.2.2
pyyaml==5.3.1
samplerate==0.1.0
skia-python==87.0
toml==0.10.2
tqdm==4.54.1
wget==3.2
